CW publisher must be "singleton"
Loggy can have multiple "instances" since it is a class. The reason is the possibility to s
specify a different __filename or differents context objects. But all the logs are sent to the same
bucket/stream (which depends on the hostname). Thus, there is the need to set the
cw publisher in a separate context, so all the classes share the same